\name{Fletcher2013b.mra.supplements}
\alias{Fletcher2013mra.agreement.cohort1}
\alias{Fletcher2013mra.agreement.exp1} 
\alias{Fletcher2013mra.agreement.exp2}
\alias{Fletcher2013mra.agreement.exp3}
\alias{Fletcher2013mra.heatmap1}
\alias{Fletcher2013mra.heatmap2}
\alias{Fletcher2013mra.consensus}

\title{
Supplementary MRA analyses to reproduce results for Fletcher et al. 2013.
}

\description{
Supplementary functions to reproduce results for Fletcher et al. 2013.
}

\usage{
 Fletcher2013mra.agreement.cohort1()
 Fletcher2013mra.agreement.exp1()
 Fletcher2013mra.agreement.exp2()
 Fletcher2013mra.agreement.exp3()
 Fletcher2013mra.heatmap1()
 Fletcher2013mra.heatmap2()
 Fletcher2013mra.consensus()
}

\value{ 
All results will be saved in the current work directory.
}

\source{
Michael NC Fletcher, Mauro AA Castro, Suet-Feung Chin, Oscar Rueda, Xin Wang, Carlos Caldas, Bruce AJ Ponder, Florian Markowetz, Kerstin B Meyer. Master regulators of FGFR2 signalling and breast cancer risk. Journal paper (submitted). 2012.
}

\author{
Mauro Castro \email{mauro.a.castro@gmail.com}
}










